// src/app/shared/model/employee.ts
export interface Employee {
    Oid: string;
    userOid: string;
    fullName: string;
    age: string;
    genderId: number;
    civilStatusId: number;
    studyLevelId: number;
    position: string;
    positionTypeId: number;
    contractTypeId: number;
    staffTypeId: number;
    workingDayTypeId: number;
    rotation: true,
    positionStartDate: Date;
    experienceYears: number;
}